Use reversible URLs provided by shrinkroute.
The separate `config/routes.js` was dropped in favour of a direct integration
of all the routes into `config/express.js`. This was necessary as shrinkroute
has to be called before `app.router` but all the routes have to be configured
after the shrinkrouter middleware has been assigned using `app.use`.